+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## [Unreleased][unreleased]
+
+## 1.3 - 2015-05-15
 ## Added
 - setns.py with setns context manager
 - args_aliases.py with dicts of possible additional arguments for many cases
@@ -95,7 +97,8 @@
 - cloning.py and process.py as first working concept
 - initial commit with necessary files
 
-[unreleased](https://github.com/Friz-zy/pyspaces/compare/v1.2.4...HEAD)
+[unreleased](https://github.com/Friz-zy/pyspaces/compare/v1.3...HEAD)
+[1.2.4-1.3](https://github.com/Friz-zy/pyspaces/compare/v1.2.4...v1.3)
 [1.2.3-1.2.4](https://github.com/Friz-zy/pyspaces/compare/v1.2.3...v1.2.4)
 [1.2.2-1.2.3](https://github.com/Friz-zy/pyspaces/compare/v1.2.2...v1.2.3)
 [1.2-1.2.2](https://github.com/Friz-zy/pyspaces/compare/v1.2...v1.2.2)